refactor(stores): absorb UserService into UserStore; delete UserServiceClient (Phase 5b / task 1.3b)
Final absorption of the entity-store-layer refactor. UserStore becomes the single
owner of the User entity: it absorbs UserServiceClient's logic (the @observable
current, ensureLoaded's cache->Get->Create->PermissionDenied recovery chain,
create/updateHome/updatePreferredLanguage write-through, resendEmailVerification,
clear, and the user_id cache helpers) MOVED byte-for-byte (only mechanical
this.userService.current -> this.current rewrites). UserStore now resolves
IUserRpcClient/ILocalStorage/IAuthService directly and keeps its guest
home/language ownership + i18n-locale mirror.

The 11 IUserService consumers (auth-callback, settings, dashboard, tickets,
welcome, auth-status, user-home-selector, change-locale, follow-reconcile-task,
user-hydration-task) now use IUserStore. ProvisionResult relocated to user-store.
UserServiceClient is deleted; the IUserRpcClient proto client is untouched.

Auth bootstrap behavior is identical: signup creates the user + persists
home/language at Create-time + publishes GuestMigrationRequested + shows
postSignup only for a genuinely-new account; sign-in + hydration backfill work.

With this, GuestService, GuestDataMergeService, FollowServiceClient, and
UserServiceClient are all dissolved into the store layer.

Refs: #368
