+++ src/components/auth-status.ts
@@ -1,10 +1,10 @@
 import { ILogger, resolve } from 'aurelia'
 import { IAuthService } from '../services/auth-service'
-import { IUserService } from '../services/user-service'
+import { IUserStore } from '../services/user-store'
 
 export class AuthStatus {
 	public readonly auth = resolve(IAuthService)
-	private readonly userService = resolve(IUserService)
+	private readonly userStore = resolve(IUserStore)
 	private readonly logger = resolve(ILogger).scopeTo('AuthStatus')
 
 	public async signIn(): Promise<void> {
@@ -24,7 +24,7 @@ export class AuthStatus {
 		// session — mirrors the pattern in settings-route.signOut(). Required
 		// by the user-account-sync spec scenario "Cached userID is cleared on
 		// sign-out".
-		this.userService.clear()
+		this.userStore.clear()
 		await this.auth.signOut()
 	}
 }
