+++ src/routes/tickets/tickets-route.ts
@@ -3,7 +3,7 @@ import QRCode from 'qrcode'
 import { ITicketRpcClient } from '../../adapter/rpc/client/ticket-client'
 import type { Ticket } from '../../entities/ticket'
 import { IProofService } from '../../services/proof-service'
-import { IUserService } from '../../services/user-service'
+import { IUserStore } from '../../services/user-store'
 
 export class TicketsRoute {
 	public tickets: Ticket[] = []
@@ -19,7 +19,7 @@ export class TicketsRoute {
 	private readonly logger = resolve(ILogger).scopeTo('TicketsRoute')
 	private readonly ticketClient = resolve(ITicketRpcClient)
 	private readonly proofService = resolve(IProofService)
-	private readonly userService = resolve(IUserService)
+	private readonly userStore = resolve(IUserStore)
 	private abortController: AbortController | null = null
 
 	public async loading(): Promise<void> {
@@ -28,7 +28,7 @@ export class TicketsRoute {
 		this.abortController = new AbortController()
 
 		try {
-			const userId = this.userService.current?.id
+			const userId = this.userStore.current?.id
 			if (!userId) {
 				this.error = 'Not signed in.'
 				return
