+++ src/util/change-locale.ts
@@ -1,6 +1,5 @@
 import type { I18N } from '@aurelia/i18n'
 import type { IAuthService } from '../services/auth-service'
-import type { IUserService } from '../services/user-service'
 import type { IUserStore } from '../services/user-store'
 
 export const SUPPORTED_LANGUAGES = ['ja', 'en'] as const
@@ -41,7 +40,6 @@ export function normalizeToSupportedLanguage(detected: string): string {
 export interface ChangeLocaleDeps {
 	readonly i18n: I18N
 	readonly auth: IAuthService
-	readonly userService: IUserService
 	readonly userStore: IUserStore
 }
 
@@ -57,8 +55,8 @@ export interface ChangeLocaleDeps {
  *   raw localStorage — keeps `UserStore.currentLanguage` reactive so the
  *   language selector highlight follows the active locale. No backend call.
  *
- * - **Authenticated** (Settings page): call `UpdatePreferredLanguage` first,
- *   apply `i18n.setLocale` after success. The backend row is the source
+ * - **Authenticated** (Settings page): call `UserStore.updatePreferredLanguage`
+ *   first, apply `i18n.setLocale` after success. The backend row is the source
  *   of truth; the hydration task clears the legacy localStorage key on
  *   next boot.
  *
@@ -69,7 +67,7 @@ export async function changeLocale(
 	deps: ChangeLocaleDeps,
 	lang: string,
 ): Promise<void> {
-	const { i18n, auth, userService, userStore } = deps
+	const { i18n, auth, userStore } = deps
 
 	if (!isSupportedLanguage(lang)) {
 		throw new TypeError(
@@ -83,6 +81,6 @@ export async function changeLocale(
 		return
 	}
 
-	await userService.updatePreferredLanguage(lang)
+	await userStore.updatePreferredLanguage(lang)
 	await i18n.setLocale(lang)
 }
