+++ test/components/auth-status.spec.ts
@@ -2,21 +2,21 @@ import { Registration } from 'aurelia'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { AuthStatus } from '../../src/components/auth-status'
 import { IAuthService } from '../../src/services/auth-service'
-import { IUserService } from '../../src/services/user-service'
+import { IUserStore } from '../../src/services/user-store'
 import { createTestContainer } from '../helpers/create-container'
 import { createMockAuth } from '../helpers/mock-auth'
 
 describe('AuthStatus', () => {
 	let sut: AuthStatus
 	let mockAuth: ReturnType<typeof createMockAuth>
-	let mockUserService: { clear: ReturnType<typeof vi.fn> }
+	let mockUserStore: { clear: ReturnType<typeof vi.fn> }
 
 	beforeEach(() => {
 		mockAuth = createMockAuth()
-		mockUserService = { clear: vi.fn() }
+		mockUserStore = { clear: vi.fn() }
 		const container = createTestContainer(
 			Registration.instance(IAuthService, mockAuth as IAuthService),
-			Registration.instance(IUserService, mockUserService as never),
+			Registration.instance(IUserStore, mockUserStore as never),
 		)
 		container.register(AuthStatus)
 		sut = container.get(AuthStatus)
@@ -52,9 +52,9 @@ describe('AuthStatus', () => {
 
 		// Assert: both called, with clear preceding signOut so the localStorage
 		// entry does not outlive the session.
-		expect(mockUserService.clear).toHaveBeenCalledOnce()
+		expect(mockUserStore.clear).toHaveBeenCalledOnce()
 		expect(mockAuth.signOut).toHaveBeenCalledOnce()
-		expect(mockUserService.clear.mock.invocationCallOrder[0]).toBeLessThan(
+		expect(mockUserStore.clear.mock.invocationCallOrder[0]).toBeLessThan(
 			(mockAuth.signOut as ReturnType<typeof vi.fn>).mock
 				.invocationCallOrder[0],
 		)
